backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "narrative-treemap-backend"
version = "0.1.0"
description = "Local-first narrative treemap backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "feedparser>=6.0.11",
  "beautifulsoup4>=4.12.3",
  "sentence-transformers>=3.0.1",
  "scikit-learn>=1.5.0",
  "numpy>=1.26.4",
  "python-dateutil>=2.9.0",
  "pydantic-settings>=2.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
